feat(carteira): suporta acoes fracionarias (quantidade decimal)
Operacao.quantidade era Integer, bloqueando qualquer fracao de acao.
Troca para BigDecimal em toda a cadeia: entidade, DTOs (Operacao e
Posicao), PosicaoCalculator (media ponderada, saldo minimo historico)
e os dois services que consomem essas comparacoes. Nenhuma mudanca de
regra de negocio -- so o tipo do numero, mesma logica de venda a
descoberto e cancelamento que ja existia continua valendo para
quantidades fracionarias.

+++ src/main/java/com/curso/gestaoinvestimentos/dto/PosicaoDTO.java
@@ -5,7 +5,7 @@ import java.math.BigDecimal;
 public record PosicaoDTO(
         String acaoTicker,
         String corretoraNome,
-        Integer quantidade,
+        BigDecimal quantidade,
         BigDecimal precoMedio,
         BigDecimal valorInvestido,
         BigDecimal valorAtual
